Allow signing external ESL at db level
We are moving to authenticating the bootable images by hashes
of the EFI binaries instead of an embedded signature. This way the ESL
with the hashes is built by the client (yocto) and sent to sign
as is. This patch extends the secureboot/db signing API to allow
sending an external ESL instead of cert_id.

This only applies to db, PK and KEK are still certificates.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>