Dockerfile: build NXP's CST tool
The tool is only available via NXP's download page prior license acceptance.
To avoid licenses issues, the source must be manually downloaded and
installed prior to building the signing server app.

The source can be fetched from:
https://www.nxp.com/webapp/Download?colCode=IMX_CST_TOOL_NEW&location=null

And it is only provided for x86_64 architectures, so the signing server
is now restricted to running on x86_64 hardware.

Change-type: minor
Signed-off-by: Alex Gonzalez <alexg@balena.io>