Add NXP i.MX PKI tree generation and CSF support
To sign NXP i.MX artifacts, a PKI tree for the specific security framework
(hab/ahab) needs to be created.

The PKI tree generation also returns an efuses hash artifact that is to
be programmed into the device's efuses to authenticate artifacts in the
secure boot chain of trust.

Artifacts are signed by providing its payload and a CSF template for the
CST tool to use. The CSF binary is then returned and can be appended to
artifacts as needed.

Signed-off-by: Alex Gonzalez <alexg@balena.io>