Use python 3.12 and update dependencies
The connexion 3 update switches from using a WSGI server to a ASGI one,
so needs to be served by gunicorn using the uvicorn backend.

Also update base images to using current stable versions:

* Debian bookworm
* Alpine 3.19
* Golang 1.22

Stop using balenalib images as they are not up to date and not likely
to keep up with upstream.

Change-type: minor
Signed-off-by: Alex Gonzalez <alexg@balena.io>